safeRiderWeb/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
requests==2.31.0
twilio==9.0.3

mapbox~=0.2.0